`run_rustc` respect `j` option
It respects desires `-j` option for make which calls it, and doesn't
owerride it by `-j 3`.

Thus, it also build `libstd.rlib` as dependency for `cargo` and `rustc`
which prevent it to build in the parallels when `run_rustc` is called on
clean tree.

And it adds one a few dependency from `cargo` and `rustc` to `minicargo`,
because before it was as side effect.

And it allows to use installed in the system `minicargo` and `mrustc` to
build everything up.